+++ UIKitSwagger/Color/GrayscaleComponents.swift
@@ -1,5 +1,5 @@
 //
-//  UIColorGrayscaleComponents.swift
+//  GrayscaleComponents.swift
 //  UIKitSwagger
 //
 //  Created by Sam Odom on 12/13/14.
@@ -11,7 +11,7 @@ import UIKit
 /**
 Convenience structure to hold the white and alpha component values of an instance of `UIColor`.
 */
-public struct UIColorGrayscaleComponents: UIColorComponents {
+public struct GrayscaleComponents: ColorComponents {
     public let white: CGFloat
     public let alpha: CGFloat
 
@@ -31,11 +31,11 @@ public struct UIColorGrayscaleComponents: UIColorComponents {
 /**
 Equatability of grayscale components.
 */
-extension UIColorGrayscaleComponents: Equatable {
+extension GrayscaleComponents: Equatable {
 
 }
 
-public func ==(lhs: UIColorGrayscaleComponents, rhs: UIColorGrayscaleComponents) -> Bool {
+public func ==(lhs: GrayscaleComponents, rhs: GrayscaleComponents) -> Bool {
     return componentValuesEqualWithinTolerance(lhs.white, rhs.white) &&
         componentValuesEqualWithinTolerance(lhs.alpha, rhs.alpha)
 }
@@ -45,12 +45,12 @@ public extension UIColor {
     /**
     Property that returns the grayscale components of the color in a structure.
     */
-    public var grayscaleComponents: UIColorGrayscaleComponents {
+    public var grayscaleComponents: GrayscaleComponents {
         var whiteValue = CGFloat(0)
         var alphaValue = CGFloat(0)
 
         getWhite(&whiteValue, alpha: &alphaValue)
-        return UIColorGrayscaleComponents(white: whiteValue, alpha: alphaValue)
+        return GrayscaleComponents(white: whiteValue, alpha: alphaValue)
     }
 
 
@@ -59,26 +59,26 @@ public extension UIColor {
 /**
 Component conversion methods.
 */
-public extension UIColorGrayscaleComponents {
+public extension GrayscaleComponents {
 
     /**
     Converts grayscale components into RGB components.
     */
-    public func asRGBComponents() -> UIColorRGBComponents {
+    public func asRGBComponents() -> RGBComponents {
         return color().rgbComponents
     }
 
     /**
     Converts grayscale components into HSB components.
     */
-    public func asHSBComponents() -> UIColorHSBComponents {
+    public func asHSBComponents() -> HSBComponents {
         return color().hsbComponents
     }
 
     /**
     Converts grayscale components into CMYK components.
     */
-    public func asCMYKComponents() -> UIColorCMYKComponents {
+    public func asCMYKComponents() -> CMYKComponents {
         return color().cmykComponents
     }