+++ UIKitSwagger/Color/HSBComponents.swift
@@ -1,5 +1,5 @@
 //
-//  UIColorHSBComponents.swift
+//  HSBComponents.swift
 //  UIKitSwagger
 //
 //  Created by Sam Odom on 12/13/14.
@@ -11,7 +11,7 @@ import UIKit
 /**
 Convenience structure to hold the hue, saturation, brightness and alpha component values of an instance of `UIColor`.
 */
-public struct UIColorHSBComponents: UIColorComponents {
+public struct HSBComponents: ColorComponents {
     public let hue: CGFloat
     public let saturation: CGFloat
     public let brightness: CGFloat
@@ -35,11 +35,11 @@ public struct UIColorHSBComponents: UIColorComponents {
 /**
 Equatability of HSB components.
 */
-extension UIColorHSBComponents: Equatable {
+extension HSBComponents: Equatable {
 
 }
 
-public func ==(lhs: UIColorHSBComponents, rhs: UIColorHSBComponents) -> Bool {
+public func ==(lhs: HSBComponents, rhs: HSBComponents) -> Bool {
     return componentValuesEqualWithinTolerance(lhs.hue, rhs.hue) &&
         componentValuesEqualWithinTolerance(lhs.saturation, rhs.saturation) &&
         componentValuesEqualWithinTolerance(lhs.brightness, rhs.brightness) &&
@@ -51,14 +51,14 @@ public extension UIColor {
     /**
     Property that returns the HSB components of the color in a structure.
     */
-    public var hsbComponents: UIColorHSBComponents {
+    public var hsbComponents: HSBComponents {
         var hueValue = CGFloat(0)
         var saturationValue = CGFloat(0)
         var brightnessValue = CGFloat(0)
         var alphaValue = CGFloat(0)
 
         getHue(&hueValue, saturation: &saturationValue, brightness: &brightnessValue, alpha: &alphaValue)
-        return UIColorHSBComponents(hue: hueValue, saturation: saturationValue, brightness: brightnessValue, alpha: alphaValue)
+        return HSBComponents(hue: hueValue, saturation: saturationValue, brightness: brightnessValue, alpha: alphaValue)
     }
 
 }
@@ -66,26 +66,26 @@ public extension UIColor {
 /**
 Component conversion methods.
 */
-public extension UIColorHSBComponents {
+public extension HSBComponents {
 
     /**
     Converts HSB components into RGB components.
     */
-    public func asRGBComponents() -> UIColorRGBComponents {
+    public func asRGBComponents() -> RGBComponents {
         return color().rgbComponents
     }
 
     /**
     Converts HSB components into grayscale components.
     */
-    public func asGrayscaleComponents() -> UIColorGrayscaleComponents {
+    public func asGrayscaleComponents() -> GrayscaleComponents {
         return color().grayscaleComponents
     }
 
     /**
     Converts HSB components into CNYK components.
     */
-    public func asCMYKComponents() -> UIColorCMYKComponents {
+    public func asCMYKComponents() -> CMYKComponents {
         return color().cmykComponents
     }