+++ UIKitSwaggerTests/Tests/CMYKComponentsTests.swift
@@ -15,7 +15,7 @@ import XCTest
 class CMYKComponentsTests: XCTestCase {
 
     var components =
-    UIColorCMYKComponents(
+    CMYKComponents(
         cyan: randomCyanValue,
         magenta: randomMagentaValue,
         yellow: randomYellowValue,
@@ -149,7 +149,7 @@ class CMYKComponentsTests: XCTestCase {
     //  MARK: - Components structure
 
     func testCMYKComponentStructureWithDefaultAlpha() {
-        components = UIColorCMYKComponents(
+        components = CMYKComponents(
             cyan: randomCyanValue,
             magenta: randomMagentaValue,
             yellow: randomYellowValue,
@@ -174,7 +174,7 @@ class CMYKComponentsTests: XCTestCase {
 
     func testEqualityOfCMYKComponentStructure() {
         let moreComponents =
-        UIColorCMYKComponents(
+        CMYKComponents(
             cyan: randomCyanValue,
             magenta: randomMagentaValue,
             yellow: randomYellowValue,
@@ -186,7 +186,7 @@ class CMYKComponentsTests: XCTestCase {
 
     func testInequalityOfCMYKComponentsWithMismatchedCyanValues() {
         let moreComponents =
-        UIColorCMYKComponents(
+        CMYKComponents(
             cyan: nudgeComponentValue(randomCyanValue),
             magenta: randomMagentaValue,
             yellow: randomYellowValue,
@@ -198,7 +198,7 @@ class CMYKComponentsTests: XCTestCase {
 
     func testInequalityOfCMYKComponentsWithMismatchedMagentaValues() {
         let moreComponents =
-        UIColorCMYKComponents(
+        CMYKComponents(
             cyan: randomCyanValue,
             magenta: nudgeComponentValue(randomMagentaValue),
             yellow: randomYellowValue, 
@@ -210,7 +210,7 @@ class CMYKComponentsTests: XCTestCase {
 
     func testInequalityOfCMYKComponentsWithMismatchedYellowValues() {
         let moreComponents =
-        UIColorCMYKComponents(
+        CMYKComponents(
             cyan: randomCyanValue,
             magenta: randomMagentaValue,
             yellow: nudgeComponentValue(randomYellowValue),
@@ -222,7 +222,7 @@ class CMYKComponentsTests: XCTestCase {
 
     func testInequalityOfCMYKComponentsWithMismatchedKeyValues() {
         let moreComponents =
-        UIColorCMYKComponents(
+        CMYKComponents(
             cyan: randomCyanValue,
             magenta: randomMagentaValue,
             yellow: randomYellowValue,
@@ -234,7 +234,7 @@ class CMYKComponentsTests: XCTestCase {
 
     func testInequalityOfCMYKComponentsWithMismatchedAlphaValues() {
         let moreComponents =
-        UIColorCMYKComponents(
+        CMYKComponents(
             cyan: randomCyanValue,
             magenta: randomMagentaValue,
             yellow: randomYellowValue,
@@ -248,19 +248,19 @@ class CMYKComponentsTests: XCTestCase {
 
     func testCMYKComponentsWithRGBColor() {
         computeExpectedCMYKValuesForColor(sampleRGBColor)
-        components = UIColorCMYKComponents(cyan: expectedCyan, magenta: expectedMagenta, yellow: expectedYellow, key: expectedKey, alpha: randomAlphaValue)
+        components = CMYKComponents(cyan: expectedCyan, magenta: expectedMagenta, yellow: expectedYellow, key: expectedKey, alpha: randomAlphaValue)
         XCTAssertEqual(sampleRGBColor.cmykComponents, components, "The cyan, magenta, yellow, key and alpha components of the color should be provided")
     }
 
     func testCMYKComponentsWithHSBColor() {
         computeExpectedCMYKValuesForColor(sampleHSBColor)
-        components = UIColorCMYKComponents(cyan: expectedCyan, magenta: expectedMagenta, yellow: expectedYellow, key: expectedKey, alpha: randomAlphaValue)
+        components = CMYKComponents(cyan: expectedCyan, magenta: expectedMagenta, yellow: expectedYellow, key: expectedKey, alpha: randomAlphaValue)
         XCTAssertEqual(sampleHSBColor.cmykComponents, components, "The cyan, magenta, yellow, key and alpha components of the color should be provided")
     }
 
     func testCMYKComponentsWithMonochromeColor() {
         expectedKey = 1 - sampleMonochromeColor.white
-        components = UIColorCMYKComponents(cyan: 0, magenta: 0, yellow: 0, key: expectedKey, alpha: randomAlphaValue)
+        components = CMYKComponents(cyan: 0, magenta: 0, yellow: 0, key: expectedKey, alpha: randomAlphaValue)
         XCTAssertEqual(sampleMonochromeColor.cmykComponents, components, "The cyan, magenta, yellow, key and alpha components of the color should be provided")
     }
 
@@ -268,49 +268,49 @@ class CMYKComponentsTests: XCTestCase {
 
     func testCMYKComponentsWithBlackColor() {
         components = Black.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 0, magenta: 0, yellow: 0, key: 1, alpha: 1)
+        let expected = CMYKComponents(cyan: 0, magenta: 0, yellow: 0, key: 1, alpha: 1)
         XCTAssertEqual(components, expected, "Black should produce the appropriate CMYK components")
     }
 
     func testCMYKComponentsWithWhiteColor() {
         components = White.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 0, magenta: 0, yellow: 0, key: 0, alpha: 1)
+        let expected = CMYKComponents(cyan: 0, magenta: 0, yellow: 0, key: 0, alpha: 1)
         XCTAssertEqual(components, expected, "White should produce the appropriate CMYK components")
     }
 
     func testCMYKComponentsWithRedColor() {
         components = Red.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 0, magenta: 1, yellow: 1, key: 0, alpha: 1)
+        let expected = CMYKComponents(cyan: 0, magenta: 1, yellow: 1, key: 0, alpha: 1)
         XCTAssertEqual(components, expected, "Red should produce the appropriate CMYK components")
     }
 
     func testCMYKComponentsWithGreenColor() {
         components = Green.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 1, magenta: 0, yellow: 1, key: 0, alpha: 1)
+        let expected = CMYKComponents(cyan: 1, magenta: 0, yellow: 1, key: 0, alpha: 1)
         XCTAssertEqual(components, expected, "Green should produce the appropriate CMYK components")
     }
 
     func testCMYKComponentsWithBlueColor() {
         components = Blue.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 1, magenta: 1, yellow: 0, key: 0, alpha: 1)
+        let expected = CMYKComponents(cyan: 1, magenta: 1, yellow: 0, key: 0, alpha: 1)
         XCTAssertEqual(components, expected, "Blue should produce the appropriate CMYK components")
     }
 
     func testCMYKComponentsWithCyanColor() {
         components = Cyan.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 1, magenta: 0, yellow: 0, key: 0, alpha: 1)
+        let expected = CMYKComponents(cyan: 1, magenta: 0, yellow: 0, key: 0, alpha: 1)
         XCTAssertEqual(components, expected, "Cyan should produce the appropriate CMYK components")
     }
 
     func testCMYKComponentsWithMagentaColor() {
         components = Magenta.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 0, magenta: 1, yellow: 0, key: 0, alpha: 1)
+        let expected = CMYKComponents(cyan: 0, magenta: 1, yellow: 0, key: 0, alpha: 1)
         XCTAssertEqual(components, expected, "Magenta should produce the appropriate CMYK components")
     }
 
     func testCMYKComponentsWithYellowColor() {
         components = Yellow.cmykComponents
-        let expected = UIColorCMYKComponents(cyan: 0, magenta: 0, yellow: 1, key: 0, alpha: 1)
+        let expected = CMYKComponents(cyan: 0, magenta: 0, yellow: 1, key: 0, alpha: 1)
         XCTAssertEqual(components, expected, "Yellow should produce the appropriate CMYK components")
     }